chore: clean up CategoryLabelWithIcon.tsx
