+++ translations/nl.json
@@ -204,6 +204,7 @@
     "details": "Details",
     "filters": "Filters",
     "show_menu": "Toon {menu_object}",
-    "close_menu": "Sluit {menu_object}"
+    "close_menu": "Sluit {menu_object}",
+    "category_icon_alt_text": "Icoon voor categorie: {category}"
   }
 }
